# hamming arm64 simd implementation