--- a/adapters/repos/db/vector/hnsw/distancer/asm/hamming_stub_arm64.go
+++ b/adapters/repos/db/vector/hnsw/distancer/asm/hamming_stub_arm64.go
@@ -18,7 +18,6 @@ package asm
 //go:generate goat ../c/hamming_arm64.c -O3 -e="-mfpu=neon-fp-armv8" -e="-mfloat-abi=hard" -e="--target=arm64" -e="-march=armv8-a+simd+fp"
 
 import (
-	"reflect"
 	"unsafe"
 )
 
@@ -89,16 +88,12 @@ func Hamming(x []float32, y []float32) float32 {
 
 	var res float32
 
-	// The C function expects pointers to the underlying array, not slices.
-	hdrx := (*reflect.SliceHeader)(unsafe.Pointer(&x))
-	hdry := (*reflect.SliceHeader)(unsafe.Pointer(&y))
-
 	l := len(x)
 	hamming(
 		// The slice header contains the address of the underlying array.
 		// We only need to cast it to a pointer.
-		unsafe.Pointer(hdrx.Data),
-		unsafe.Pointer(hdry.Data),
+		unsafe.Pointer(unsafe.SliceData(x)),
+		unsafe.Pointer(unsafe.SliceData(y)),
 		// The C function expects pointers to the result and the length of the arrays.
 		unsafe.Pointer(&res),
 		unsafe.Pointer(&l))