--- a/adapters/repos/db/vector/hnsw/distancer/bench_arm64_test.go
+++ b/adapters/repos/db/vector/hnsw/distancer/bench_arm64_test.go
@@ -60,7 +60,6 @@ func BenchmarkDot(b *testing.B) {
 	}
 }
 
-
 func benchmarkHammingGo(b *testing.B, dims int) {
 	r := getRandomSeed()