--- a/adapters/repos/db/vector/hnsw/distancer/hamming.go
+++ b/adapters/repos/db/vector/hnsw/distancer/hamming.go
@@ -40,6 +40,17 @@ func (l Hamming) Distance(b []float32) (float32, bool, error) {
 	return hammingImpl(l.a, b), true, nil
 }
 
+func HammingDistanceGo(a, b []float32) float32 {
+	var sum float32
+	for i := range a {
+		if a[i] != b[i] {
+			sum += float32(1)
+		}
+	}
+	return sum
+}
+
+
 type HammingProvider struct{}
 
 func NewHammingProvider() HammingProvider {