--- a/adapters/repos/db/vector/hnsw/distancer/hamming_test.go
+++ b/adapters/repos/db/vector/hnsw/distancer/hamming_test.go
@@ -163,25 +163,16 @@ func TestCompareHammingDistanceImplementations(t *testing.T) {
 			}
 
 			for i := 0; i < count; i++ {
-
 				res, ok, err := NewHammingProvider().New(vec1s[i]).Distance(vec2s[i])
-				if err != nil {
-					panic(err)
-				}
 
-				if !ok {
-					panic("not ok")
-				}
+				require.NoError(t, err)
+				require.True(t, ok)
 
 				resControl := HammingDistanceGo(vec1s[i], vec2s[i])
 
-				delta := float64(0.01)
-				diff := float64(resControl) - float64(res)
-				if diff < -delta || diff > delta {
+				if resControl != res {
 					countFailed++
-
-					fmt.Printf("run %d: match: %f != %f, %d\n", i, resControl, res, (unsafe.Pointer(&vec1s[i][0])))
-
+					t.Fatalf("run %d: match: %f != %f, %d\n", i, resControl, res, (unsafe.Pointer(&vec1s[i][0])))
 					t.Fail()
 				}