### What's being changed:

SIMD implementation for hamming distance for ARM64

Benchmark:
```
goos: darwin
goarch: arm64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer
                          │ pure_go.txt  │           arm64_simd.txt            │
                          │    sec/op    │    sec/op     vs base               │
Hamming/30_dimensions-8     10.725n ± 1%   7.297n ±  3%  -31.96% (p=0.002 n=6)
Hamming/32_dimensions-8     11.340n ± 0%   5.938n ±  6%  -47.64% (p=0.002 n=6)
Hamming/128_dimensions-8     48.73n ± 1%   11.46n ±  5%  -76.47% (p=0.002 n=6)
Hamming/256_dimensions-8     89.76n ± 1%   19.12n ±  9%  -78.70% (p=0.002 n=6)
Hamming/300_dimensions-8    103.60n ± 0%   21.86n ±  2%  -78.90% (p=0.002 n=6)
Hamming/384_dimensions-8    130.20n ± 0%   26.72n ± 13%  -79.48% (p=0.002 n=6)
Hamming/600_dimensions-8    198.60n ± 1%   40.02n ±  1%  -79.85% (p=0.002 n=6)
Hamming/768_dimensions-8    251.25n ± 0%   49.64n ±  1%  -80.24% (p=0.002 n=6)
Hamming/1024_dimensions-8   332.65n ± 0%   64.96n ±  0%  -80.47% (p=0.002 n=6)
geomean                      78.94n        20.76n        -73.70%
```

### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [x] Chaos pipeline run or not necessary. Link to pipeline:
- [x] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.