@ajit283 2024-04-10
updated benchmark:
```
goos: darwin
goarch: arm64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer
                          │  pure_go.txt  │            arm64_neon.txt             │
                          │    sec/op     │    sec/op      vs base                │
Hamming/1_dimensions-8      0.9368n ±  0%   4.4415n ±  2%  +374.14% (p=0.002 n=6)
Hamming/2_dimensions-8       1.879n ±  1%    2.068n ±  2%   +10.06% (p=0.002 n=6)
Hamming/3_dimensions-8       2.190n ±  1%    6.354n ±  1%  +190.16% (p=0.002 n=6)
Hamming/4_dimensions-8       2.502n ±  1%    2.551n ±  1%    +1.94% (p=0.009 n=6)
Hamming/5_dimensions-8       2.811n ±  1%    4.592n ±  6%   +63.38% (p=0.002 n=6)
Hamming/6_dimensions-8       3.121n ±  1%    4.143n ± 15%   +32.75% (p=0.002 n=6)
Hamming/8_dimensions-8       3.751n ±  1%    4.176n ±  5%   +11.34% (p=0.002 n=6)
Hamming/10_dimensions-8      4.364n ±  0%    4.773n ±  6%    +9.38% (p=0.002 n=6)
Hamming/12_dimensions-8      4.993n ±  0%    5.424n ±  6%    +8.62% (p=0.002 n=6)
Hamming/16_dimensions-8      6.253n ±  1%    4.752n ±  1%   -24.01% (p=0.002 n=6)
Hamming/24_dimensions-8      8.749n ±  0%    5.529n ±  3%   -36.81% (p=0.002 n=6)
Hamming/30_dimensions-8     10.600n ±  0%    7.708n ±  4%   -27.29% (p=0.002 n=6)
Hamming/31_dimensions-8     10.935n ±  0%    7.994n ± 12%   -26.90% (p=0.002 n=6)
Hamming/32_dimensions-8     11.225n ±  0%    5.782n ±  2%   -48.49% (p=0.002 n=6)
Hamming/64_dimensions-8     21.240n ±  0%    7.655n ±  1%   -63.96% (p=0.002 n=6)
Hamming/67_dimensions-8     22.175n ±  0%    8.875n ±  2%   -59.98% (p=0.002 n=6)
Hamming/128_dimensions-8     48.34n ±  1%    11.50n ±  1%   -76.22% (p=0.002 n=6)
Hamming/256_dimensions-8     89.05n ±  1%    19.14n ±  1%   -78.50% (p=0.002 n=6)
Hamming/260_dimensions-8     90.38n ±  0%    19.45n ±  1%   -78.48% (p=0.002 n=6)
Hamming/299_dimensions-8    103.95n ±  1%    22.87n ±  1%   -78.00% (p=0.002 n=6)
Hamming/300_dimensions-8    103.50n ±  1%    22.08n ±  4%   -78.67% (p=0.002 n=6)
Hamming/384_dimensions-8    128.90n ±  1%    26.94n ±  4%   -79.10% (p=0.002 n=6)
Hamming/390_dimensions-8    132.10n ±  1%    28.10n ±  2%   -78.73% (p=0.002 n=6)
Hamming/600_dimensions-8    207.75n ± 14%    39.88n ±  3%   -80.80% (p=0.002 n=6)
Hamming/768_dimensions-8    249.70n ±  1%    50.05n ±  6%   -79.96% (p=0.002 n=6)
Hamming/777_dimensions-8    253.05n ±  1%    52.74n ±  4%   -79.16% (p=0.002 n=6)
Hamming/784_dimensions-8    256.65n ±  1%    50.91n ±  1%   -80.17% (p=0.002 n=6)
Hamming/1024_dimensions-8   329.55n ±  1%    65.21n ±  6%   -80.21% (p=0.002 n=6)
Hamming/1536_dimensions-8   493.10n ±  0%    96.03n ±  4%   -80.53% (p=0.002 n=6)
geomean                      23.87n          11.75n         -50.78%
```